pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "edgegate"
version = "0.1.0"
description = "Edge GenAI Regression Gates for Snapdragon (AI Hub-Orchestrated)"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "EdgeGate Team", email = "team@edgegate.io"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # API Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    
    # Database
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    
    # Task Queue
    "celery>=5.3.6",
    "redis>=5.0.0",
    
    # Object Storage
    "boto3>=1.34.0",
    
    # Security
    "pynacl>=1.5.0",
    "cryptography>=41.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # JSON Schema Validation
    "jsonschema>=4.21.0",
    
    # JSONPath for metric extraction
    "jsonpath-ng>=1.6.0",
    
    # AI Hub SDK (optional, for real integration)
    # "qai_hub",
    
    # PyTorch for probe model generation
    "torch>=2.1.0",
    
    # ONNX for model conversion
    "onnx>=1.15.0",
    
    # HTTP client
    "httpx>=0.26.0",
    
    # Utils
    "python-multipart>=0.0.6",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-python-dateutil>=2.8.19",
    "types-passlib>=1.7.7",
    "types-redis>=4.6.0",
]

aihub = [
    "qai_hub>=0.15.0",
]

[project.scripts]
edgegate = "edgegate.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["edgegate"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "aihub: marks tests requiring AI Hub credentials (skip without QAIHUB_API_TOKEN)",
    "integration: marks integration tests requiring docker services",
    "security: marks security-focused tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
